同上

+++ dxg/common/parser.h
@@ -72,12 +72,12 @@ namespace daxia
 #pragma pack(1)
 #endif
 				// 数据包头
-					struct ATTRIBUTE_PACKED PacketHead
+				struct ATTRIBUTE_PACKED PacketHead
 				{
 					byte	magic;			// 恒定为88
 					byte	hearbeat;		// 心跳包标识
 					unsigned int len;		// 数据长度，不包括本包头
-					unsigned long reserve;	// 保留数据
+					unsigned int reserve;	// 保留数据
 				};
 
 #ifdef _MSC_VER
